backend-services/services/hand-smart-order-router/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-cloud-pubsub==2.18.4
google-cloud-storage==2.10.0
google-cloud-bigquery==3.13.0
google-cloud-bigtable==2.21.0
google-cloud-secretmanager==2.16.3
google-cloud-logging==3.8.0
psycopg2-binary==2.9.7
redis==4.6.0
web3==6.11.1
ccxt==4.0.0
